Add `runGradleBuild` helper function
Also:
 * Add popular Gradle-related constants.
